% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHS.monthly.R
\docType{data}
\name{DHS3.monthly}
\alias{DHS3.monthly}
\title{Daniel-Hirshleifer-Sun Three-Factors Data Set}
\format{
An \code{xts} object containing observations of Daniel-Hirshleifer-Sun (2020)
three-factors data set on the U.S. Stock Market.
\itemize{
\item \strong{Frequency}: Monthly.
\item \strong{Date Range}: 1972-07 to 2018-12.
\item \strong{Data updated}: 2020-08-25 20:48:48 CEST.
\item \strong{RF}: A numeric. The risk-free rate on 1-month U.S. T-Bill. See 'RF variable' section below.
\item \strong{MKT.RF}: A numeric. The market portfolio proxy return net of risk-free rate factor. See 'MKT.RF factor' section below.
\item \strong{PEAD}: A numeric. The \emph{post-earnings announcement drift} behavioral mispricing factor. See 'PEAD factor' section below.
\item \strong{FIN}: A numeric. The \emph{financing} behavioral mispricing factor. See 'FIN factor' section below.
}

The object consists of 558 rows and 2 columns.
}
\source{
\url{http://www.kentdaniel.net/data/DHS_factors.xlsx}
}
\usage{
data("DHS3.monthly")
}
\description{
\code{DHS3.monthly} is the Daniel-Hirshleifer-Sun (2020) three-factors monthly data
series on U.S. stock market from 1972-07 to 2018-12.
}
\details{
In addition to column definitions, this section contains a glimpse into factors
construction and their underlying variables.
}
\section{Daniel-Hirshleifer-Sun Three-Factors Construction}{


Daniel-Hirshleifer-Sun (2020) factors construction is the following procedure:
\itemize{
\item First, all NYSE, AMEX, and NASDAQ common stocks (CRSP 10 or 11 share codes,
excluding financial firms and firms with negative book equity).
\item Second, at June end firms are assigned to one of two size groups ("small" and "big"),
depending on their ME being below or above the NYSE median size breakpoint.
\item Finally, firms are also independently sorted into one of three financing groups
("low", "middle", "high") based on: either stocks' 1-year NSI and 5-year CSI
financing measures rankings for the \emph{FIN} factor, or the 4-day cumulative
abnormal return around the most recent quarterly earnings announcement date
(\emph{CAR}) for the \emph{PEAD} factor. Both sorts are with respect to NYSE 20th and
80th percentiles breakpoints.
}
}

\section{RF variable}{


The \code{RF} variable refers to the \emph{risk-free rate}. It depends on the period been
considered and on the country. For example, for U.S. monthly data series is the
one month \emph{T-Bill} return.
The \code{RF} data series distributed by K. R. French with the Fama-French factors
data are usually obtained from \emph{Ibbotson Associates Inc. (Morningstar)}.
}

\section{MKT.RF factor}{


With \code{MKT.RF} we indicate the excess return on the market portfolio return proxy,
net of the risk-free rate \eqn{RF} calculated on the same period \eqn{t}, that is

\deqn{MKT.RF = MKT - RF}

or, as it is also commonly denoted in the literature,

\deqn{MKT.RF = R_{m} - R_{f}}

\eqn{MKT} is obtained by Fama-French as the value-weight return of all CRSP
firms that are incorporated in the U.S. and listed on the NYSE, AMEX, or NASDAQ
securities markets. These firms must have a CRSP share code of 10 or 11, good
shares and price data, at the beginning of the period.
}

\section{PEAD factor}{


The \emph{post-earnings announcement drift} (PEAD) factor is based on \eqn{2x3} sort
on size and earning-announcement returns, with value-weighted portfolios.

The "earning surprise" is the 4-day cumulative abnormal return around the most
recent quarterly earnings announcement date (CAR).

Thus, the monthly PEAD factor return is the arithmetic average return of the
"high earnings surprise portfolios" minus the arithmetic average return of the
"low earnings surprise portfolios".
}

\section{FIN factor}{


The \emph{financing factor} (FIN) s based on \eqn{2x3} sort on size and two financing
measures (1-year net share issuance, NSI, and 5-year composite share issuance, CSI)
rankings returns, with value-weighted portfolios.

Thus, the monthly FIN factor return is the aritmetic average return of the "low
financing portfolios" minus the arithmetic average return of the "high financing
portfolios".
}

\examples{
data(DHS3.monthly)

head(DHS3.monthly)

}
\references{
Daniel, K. and Hirshleifer, D. and Sun, L. (2020). \emph{Short-and long-horizon behavioral factors}. The Review of Financial Studies.
}
\keyword{datasets}
