\name{ExpectedReturns-package}
\alias{ExpectedReturns-package}
\alias{ExpectedReturns}
\docType{package}
\title{\packageTitle{ExpectedReturns}}
\description{\packageDescription{ExpectedReturns}}
\section{Package Content}{\packageIndices{ExpectedReturns}}
\author{\packageAuthor{ExpectedReturns}}
\section{Maintainer}{\packageMaintainer{ExpectedReturns}}
\keyword{package}
