% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q-Factors.R
\docType{data}
\name{Q5.monthly}
\alias{Q5.monthly}
\title{Hou-Xue-Zhang q-factor Five-Factors Data Set}
\format{
An \code{xts} object containing observations of Hou-Mo-Xue-Zhang q-factor five-factors
data set on U.S. Stock Market, and the risk-free rate on 1-month U.S. T-Bill.
\itemize{
\item \strong{Frequency}: Monthly.
\item \strong{Date Range}: 1967-01 to 2019-12.
\item \strong{Data updated}: 2020-08-20 19:19:20 CEST.
\item \strong{RF}: A numeric. The risk-free rate on 1-month U.S. T-Bill. See 'RF variable' section below.
\item \strong{MKT.RF}: A numeric. The market portfolio proxy return net of risk-free rate factor. See 'MKT.RF factor' section below.
\item \strong{ME}: A numeric. The size factor. See 'ME factor' section below.
\item \strong{IA}: A numeric. The investment factor. See 'IA factor' section below.
\item \strong{ROE}: A numeric. The profitability factor. See 'ROE factor' section below.
\item \strong{EG}: A numeric. The "Expected Growth" factor. See 'EG factor' section below.
}

The object consists of 636 rows and 6 columns.
}
\source{
\href{http://global-q.org/index.html}{global-q.org},
\href{http://global-q.org/uploads/1/2/2/6/122679606/factorstd_2020july.pdf}{Technical Document: Factors}
}
\usage{
data("Q5.monthly")
}
\description{
\code{Q5.monthly} is the Hou-Xue-Zhang q-factor five-factors monthly data series on
U.S. stock market from 1967-01 to 2019-12. The data set also includes the risk-free
rate on 1-month U.S. T-Bill during the same period.
}
\details{
In addition to column definitions, this section contains a glimpse into factors
construction and their underlying variables.
}
\section{Q-Factors construction}{


Authors construct their size, investment, and Roe factors from independent,
triple \eqn{2x3x3} portfolio sorts on size, investment-to-assets (I/A), and Roe.

Taking the intersection of the two size, three I/A, and three Roe groups, they
form eighteen portfolios. Monthly value-weighted portfolio returns are calculated
for the current month, and the portfolios are rebalanced monthly.
}

\section{RF variable}{


The \code{RF} variable refers to the \emph{risk-free rate}. It depends on the period been
considered and on the country. For example, for U.S. monthly data series is the
one month \emph{T-Bill} return.
The \code{RF} data series distributed by K. R. French with the Fama-French factors
data are usually obtained from \emph{Ibbotson Associates Inc. (Morningstar)}.
}

\section{BE variable}{


\emph{Book equity}, denoted \eqn{BE}, is a firms' financial variable that Fama-French
construct from Compustat data or is collected from the \emph{Moody's Industrial, Financial, and Utilities manuals}.
\eqn{BE} is the book value of stockholders' equity, plus balance sheet deferred
taxes and investment tax credit (if available), minus the book value of preferred
stock. Depending on availability, they use the redemption, liquidation, or par
value (in that order) to estimate the book value of preferred stock.
Stockholders' equity could be reported by Moody's or Compustat, otherwise the
authors measure stockholders' equity as the book value of common equity plus
the par value of preferred stock, or the book value of assets minus total
liabilities (in that order).
}

\section{Market Equity Variable}{


\emph{Market equity}, also referred to as \emph{size} and denoted \eqn{ME}, is a firms'
financial variable calculated as the product of price times shares outstanding.
The \emph{stock price} is from CRSP, whereas \emph{shares outstanding} are from Compustat
(if available) or CRSP.
}

\section{Investment-to-assets variable}{


\emph{Investment-to-assets} (I/A) is the annual change in total assets divided by
one-year-lagged total assets.
}

\section{Return on Equity variable}{


The \emph{return on equity} ratio (ROE) authors consider is the income before extraordinary
items divided by book equity (BE) lagged by a quarter.
}

\section{MKT.RF factor}{


With \code{MKT.RF} we indicate the excess return on the market portfolio return proxy,
net of the risk-free rate \eqn{RF} calculated on the same period \eqn{t}, that is

\deqn{MKT.RF = MKT - RF}

or, as it is also commonly denoted in the literature,

\deqn{MKT.RF = R_{m} - R_{f}}

\eqn{MKT} is obtained by Fama-French as the value-weight return of all CRSP
firms that are incorporated in the U.S. and listed on the NYSE, AMEX, or NASDAQ
securities markets. These firms must have a CRSP share code of 10 or 11, good
shares and price data, at the beginning of the period.
}

\section{ME factor}{


The \emph{size factor} (ME) is the small-minus-big monthly difference between the
arithmetic average of the returns on 9 small size portfolios and the arithmetic
average of the returns on 9 big size portfolios.
}

\section{IA factor}{


The \emph{investment factor} (IA) is the low-minus-high monthly difference, between
the arithmetic average of the returns on 6 low I/A portfolios and the arithmetic
average of the returns on 6 high I/A portfolios.
}

\section{ROE factor}{


The ROE factor is the high- minus-low monthly difference between the arithmetic
average of the returns on 6 high ROE portfolios and the arithmetic average of
the returns on the 6 low ROE portfolios.
}

\section{EG factor}{


The \emph{expected growth} (EG) factor is constructed from independent \eqn{2x3}
sorts on size and the expected one-year-ahead investment-to-assets change,
\eqn{E_{t}[d I/A]}. In turn, \eqn{E_{t}[d I/A]} is formed using Tobin's q,
operating cash flow-to-assets, and the change in ROE (measured as the year on
year difference in ROE).
}

\examples{
data(Q5.monthly)

head(Q5.monthly)

}
\references{
Davis, Fama, and French (2000). \emph{Characteristics, Covariances, and Average Returns: 1929-1997}. Journal of Finance.

Hou, K. and Mo, H. and Xue, C. and Zhang, L. (2019). \emph{Which factors?}. Review of Finance.

Hou, K. and Mo, H. and Xue, C. and Zhang, L. (2020). \emph{An augmented q-factor model with expected growth}. Review of Finance.
}
\seealso{
The series was generated with \code{\link[=GetFactors]{GetFactors()}}.
}
\keyword{datasets}
