% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalStats.R
\name{SignalStats}
\alias{SignalStats}
\title{Compute Trading Signals Statistics}
\usage{
SignalStats(X, symbols, signals, method, ...)
}
\arguments{
\item{X}{A list of \code{xts} objects storing signals series. See 'Details'.}

\item{symbols}{A character vector providing the names of symbols signals were computed for.}

\item{signals}{A character vector specifying the names of signals provided in \code{X}.}

\item{method}{A string indicating which stats to compute for \code{X} signals.}

\item{...}{Any other pass through parameter.}
}
\description{
Trading signals can be further analyzed with several statistics, across instruments.
Signal statistics currently implemented include:
\itemize{
\item \strong{Activity matrix} (\code{method="activity"}).
Represents the number of long or short positions over the periods considered.
It is expressed in percentage units.
\item \strong{Correlation matrix} (\code{method="cor"}).
The average correlation among signals and across instruments.
}
}
\note{
Some signals may result into an inactive position. Such positions are not considered
in the activity matrix and thus values do not sum up to unity. Whereas, they
are accounted for in the agreement matrix.
}
\references{
Baltas, A. N. and Kosowski, R. (2012). \emph{Improving time-series momentum strategies: The role of trading signals and volatility estimators}.
\href{https://risk.edhec.edu/publications/improving-time-series-momentum-strategies-role-trading-signals-and-volatility}{EDHEC-Risk Institute}.
}
\seealso{
\code{\link[=MomSignal]{MomSignal()}}
}
\author{
Vito Lestingi
}
