% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFactors.R
\name{GetFactors}
\alias{GetFactors}
\title{Scraping Academic Financial Data Libraries

This is a convenience function to provide an easy way to download research
financial data sets from several accredited sources.
Specifically, it allows to import them in the `R` environment as `xts` objects.

The scientific libraries here contemplated are from several researchers as,
unfortunately, an open source systematic collection is not available at this
time. On one hand, these resources - of admirable effort - lack important features
(financial) databases guarantee; most notably, this results in the impossibility
to leverage APIs as in most cases there aren't any.
On the other hand, these data sets satisfy important characteristics: they are
open sourced, public and free to use, collected by world renewed Researchers
of the field.

TODO: provide a general overview of data made available

Please be considerate while using the function. In particular, do not send too
frequent requests to the web source as the supported/tolerated request rate is
unknown. Usage remains at your own discretion and responsibility.
TODO: deal with loops internally so to include some sleeping time?

All the credits for collecting, maintaining and sharing data belong to the
mentioned authors or copyright holders (see 'sources' below).}
\source{
[Kenneth F. French's data library](http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html),
[Robert Shiller's online data](http://www.econ.yale.edu/~shiller/data.htm),
[AQR data sets](https://www.aqr.com/Insights/Datasets)
}
\usage{
GetFactors(x, src = "FF", freq = "monthly", term = "ST", verbose = FALSE)
}
\arguments{
\item{x}{A character specifying the query specifying the data set to download. One of 'FF3', 'FF5', 'MOM' or 'REV'. See 'Details'.}

\item{src}{A character, the source to download data from. Currently only "FF" (default) is available.}

\item{freq}{A character, specifying . One of 'annual', 'monthly', 'weekly' or 'daily'. See 'Details'.}

\item{term}{A character, to provide when `x='REV'`. Either 'ST' (short-term, default) or 'LT' (long term).}

\item{verbose}{A boolean, whether or not to retrieve additional info on data sets being downloaded.}
}
\value{
An `xts` object with columns being the requested factors.
}
\description{
Scraping Academic Financial Data Libraries

This is a convenience function to provide an easy way to download research
financial data sets from several accredited sources.
Specifically, it allows to import them in the `R` environment as `xts` objects.

The scientific libraries here contemplated are from several researchers as,
unfortunately, an open source systematic collection is not available at this
time. On one hand, these resources - of admirable effort - lack important features
(financial) databases guarantee; most notably, this results in the impossibility
to leverage APIs as in most cases there aren't any.
On the other hand, these data sets satisfy important characteristics: they are
open sourced, public and free to use, collected by world renewed Researchers
of the field.

TODO: provide a general overview of data made available

Please be considerate while using the function. In particular, do not send too
frequent requests to the web source as the supported/tolerated request rate is
unknown. Usage remains at your own discretion and responsibility.
TODO: deal with loops internally so to include some sleeping time?

All the credits for collecting, maintaining and sharing data belong to the
mentioned authors or copyright holders (see 'sources' below).
}
\details{
With respect to the Fama-French's library, the parameter `x` specifies the
factors required, which can be the Fama-French Three factor model (`'FF3'`),
Fama-French Five factor model (`'FF5'`), The Momentum factor model (`'MOM'`) or
The Reversal model (`'REV'`).

When `x` is used as "y/z" it provides additional specifications
on the data set to be queried with the syntax "model/country" (position matters).
Countries made available are 'Asia Pacific ex Japan', 'Developed', Developed ex US',
'Emerging', 'Europe', 'Japan', 'North America' and 'United States'.

Also, note that not all the Fama-French's series are available with all the
possible frequencies of `freq`.
}
\examples{
GetFactors('FF3', src='FF')

}
\author{
Vito Lestingi
}
