% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomSignal.R
\name{MomSignal}
\alias{MomSignal}
\title{Momentum Trading Signal}
\usage{
MomSignal(X, lookback, signal)
}
\arguments{
\item{X}{A list of \code{xts} objects, storing assets data. See 'Details'.}

\item{lookback}{A numeric, indicating the number of lookback periods in months.}

\item{signal}{A character, selecting the momentum signal. One of \code{SIGN}, \code{MA}, \code{EEMD}, \code{TREND}, or \code{SMT}.}
}
\value{
A list of \code{xts} objects, consisting of the chosen momentum \code{signal} for the corresponding assets data \code{X} provided.
}
\description{
Function to compute several momentum trading signals.
}
\details{
Data strictly needed in \code{X} depends on the \code{signal} chosen. For \code{SIGN} only
returns are needed. \code{MA}, \code{EEMD}, \code{TREND}, and \code{SMT} require closing prices.
}
\references{
Baltas, Akindynos-Nikolaos and Kosowski, Robert (2012). \emph{Improving time-series momentum strategies: The role of trading signals and volatility estimators}.
\href{https://risk.edhec.edu/publications/improving-time-series-momentum-strategies-role-trading-signals-and-volatility}{EDHEC-Risk Institute}
}
\author{
Vito Lestingi
}
